requirements.txt
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.27
typing_extensions==4.10.0
Werkzeug==3.0.1
